+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 Features
 - Adds support for custom ordering of columns (https://github.com/varvet/godmin/pull/168)
+- Adds passing of html options to association helper (https://github.com/varvet/godmin/pull/176)
 
 Bug fixes:
 - Fixes an issue with the template resolver and Rails 4.2.5.1 (https://github.com/varvet/godmin/pull/175)